+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## mm/dd/2021
 
 1. [Joomla](#joomla)
-   1. [](#bugfix)
+   1. [](#improved)
+      - Themes: Pass Joomla document to Gantry when rendering php template files
+   2. [](#bugfix)
       - Fixed some SEF URLs being broken/ugly (#2935)
       - Fixed `Identifier "errors" is not defined` instead of error page (#2936)
       - Fixed fatal error in error page if `Modern routing` is being used (#2924)